Refactored the ManifestContentAppDefinition to have a factory and a seperate data class.
