+++ src/Umbraco.Web/ContentApps/ContentAppDefinitionCollectionBuilder.cs
@@ -5,14 +5,16 @@ using Umbraco.Core.Composing;
 using Umbraco.Core.Logging;
 using Umbraco.Core.Manifest;
 using Umbraco.Core.Models.ContentEditing;
+using Umbraco.Core.Services;
 
 namespace Umbraco.Web.ContentApps
 {
-    public class ContentAppDefinitionCollectionBuilder : OrderedCollectionBuilderBase<ContentAppDefinitionCollectionBuilder, ContentAppDefinitionCollection, IContentAppDefinition>
+    public class ContentAppDefinitionCollectionBuilder : OrderedCollectionBuilderBase<ContentAppDefinitionCollectionBuilder, ContentAppDefinitionCollection, IContentAppFactory>
     {
         public ContentAppDefinitionCollectionBuilder(IServiceContainer container)
             : base(container)
-        { }
+        {
+        }
 
         protected override ContentAppDefinitionCollectionBuilder This => this;
 
@@ -25,14 +27,14 @@ namespace Umbraco.Web.ContentApps
             return new ContentAppDefinitionCollection(CreateItems(), logger);
         }
 
-        protected override IEnumerable<IContentAppDefinition> CreateItems(params object[] args)
+        protected override IEnumerable<IContentAppFactory> CreateItems(params object[] args)
         {
             // get the manifest parser just-in-time - injecting it in the ctor would mean that
             // simply getting the builder in order to configure the collection, would require
             // its dependencies too, and that can create cycles or other oddities
             var manifestParser = Container.GetInstance<ManifestParser>();
 
-            return base.CreateItems(args).Concat(manifestParser.Manifest.ContentApps);
+            return base.CreateItems(args).Concat(manifestParser.Manifest.ContentApps.Select(x=>new ManifestContentAppFactory(x)));
         }
     }
 }
